+++ src/tables.rs
@@ -1,5 +1,4 @@
 mod access_token;
-mod account_policy;
 mod accounts;
 mod agent;
 mod allow_network;
@@ -41,7 +40,6 @@ use rocksdb::Direction;
 use serde::{Deserialize, Serialize};
 
 pub use self::access_token::AccessToken;
-pub use self::account_policy::AccountPolicy;
 pub use self::agent::{Agent, AgentKind};
 pub use self::allow_network::{AllowNetwork, Update as AllowNetworkUpdate};
 pub use self::block_network::{BlockNetwork, Update as BlockNetworkUpdate};
@@ -201,12 +199,6 @@ impl StateDb {
     pub(crate) fn accounts(&self) -> Table<'_, Account> {
         let inner = self.inner.as_ref().expect("database must be open");
         Table::<Account>::open(inner).expect("{ACCOUNTS} table must be present")
-    }
-
-    #[must_use]
-    pub(crate) fn account_policy(&self) -> Table<'_, AccountPolicy> {
-        let inner = self.inner.as_ref().expect("database must be open");
-        Table::<AccountPolicy>::open(inner).expect("{ACCOUNT_POLICY} table must be present")
     }
 
     #[must_use]